renamed renderUserPage to renderPages
