+++ frontend/public/javascripts/renderPages.js
